+++ Formula/gomoufox.rb
@@ -1,13 +1,13 @@
 class Gomoufox < Formula
   desc "Go driver, CLI, and MCP server for Camoufox"
   homepage "https://github.com/ehmo/gomoufox"
-  version "0.1.17"
+  version "0.1.18"
   license "MIT"
 
   on_macos do
     if Hardware::CPU.arm?
-      url "https://github.com/ehmo/gomoufox/releases/download/v0.1.17/gomoufox_0.1.17_darwin_arm64.tar.gz"
-      sha256 "eed471234ffb46314fe69423c1b277cc441d7bb3483e7fcebc9bf380570e9118"
+      url "https://github.com/ehmo/gomoufox/releases/download/v0.1.18/gomoufox_0.1.18_darwin_arm64.tar.gz"
+      sha256 "908e5b9da91165e97b71afd052f6cbcbbee3cbf3e056371ee7c0fb3a6ec74e6b"
     else
       odie "gomoufox Homebrew requires Apple Silicon because pinned Camoufox has no supported macOS Intel browser binary"
     end
@@ -17,8 +17,8 @@ class Gomoufox < Formula
     if Hardware::CPU.arm?
       odie "gomoufox Homebrew requires Linux amd64 because pinned Camoufox has no supported Linux ARM browser binary"
     else
-      url "https://github.com/ehmo/gomoufox/releases/download/v0.1.17/gomoufox_0.1.17_linux_amd64.tar.gz"
-      sha256 "accc8e91eb9fe6f95865ffe6211b28e755bb1a82e5194570d177764adb2c4e1f"
+      url "https://github.com/ehmo/gomoufox/releases/download/v0.1.18/gomoufox_0.1.18_linux_amd64.tar.gz"
+      sha256 "53944059748f26704b1f1d95813ae5145812474a15b7f22ef90041fbc915dfd3"
     end
   end